// langwatch/src/server/scenarios/execution/types.ts
export type SpanQueryFn = (params: {
  projectId: string;
  traceId: string;
}) => Promise<Span[]>;